Added a server update/downloading the latest inventory from the internet
Switched over to Linux and got curl to start working.

Tried a bunch of ways to do cross-platform screen clears and ended up going with the evil system calls. Might change that in the future depending on how ambitious I am.

Code looks ugly right now, I messed around with it a lot changing a bunch of stuff around, will clean it up in the future.

There are a few redundancies in the code, so I'm going to clear that up aswell.