Now saves the displays the file from the internet AND saves it locally
The program wasn't saving the file locally before, it was just displaying whatever text from the file it was pulling off the server. This obviously creates complications when you don't have access to the internet. Everything is really messy right now as I had to figure out how to do this by hacking together a couple different tutorials from separate sources. 

Future: going to add an option to check the last downloaded file if you're not connected to the internet.